(From Sierra) Revert "STK_REBALANCE: Fix warnings"
This reverts commit a6f50a05cecd3879cd18b7cf4d90c6acd692a3bd.

Had wrong compiler in PATH

+++ packages/stk/stk_rebalance/stk_rebalance/ZoltanPartition.cpp
@@ -1,3 +1,4 @@
+
 #include <vector>
 #include <string>
 #include <sstream>
@@ -1066,12 +1067,12 @@ int  Zoltan::evaluate( int    print_stats,
   ZOLTAN_GRAPH_EVAL   graph = {{0}};
   if (Zoltan_LB_Eval_Balance( zoltan_id, print_stats, &eval)) ierr = 1;
   if (Zoltan_LB_Eval_Graph( zoltan_id, print_stats, &graph) ) ierr = 1;
-  *nobj         = static_cast<int>(eval.nobj[0]);
+  *nobj         = eval.nobj[0];
   *obj_wgt      = eval.obj_wgt[0];
-  *ncuts        = static_cast<int>(graph.cuts[0]);
+  *ncuts        = graph.cuts[0];
   *cut_wgt      = graph.cut_wgt[0];
-  *nboundary    = static_cast<int>(graph.num_boundary[0]);
-  *nadj         = static_cast<int>(graph.nnborparts[0]);
+  *nboundary    = graph.num_boundary[0];
+  *nadj         = graph.nnborparts[0];
 
   return ierr;
